requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.61.1
python-dotenv==1.0.1
pypdf==5.1.0
python-docx==1.1.2
